improve

+++ fullupdate.sh
@@ -3,25 +3,35 @@
 # and install
 # this assumes that you have the keys and coordinates for the data stored in an
 echo -n "rmc "
-source update.sh
-source ./RMCLoader/credentials.py
 
-update $rmc_location $rmc_name 
+loader="rmcloader"
+
+source update.sh
+source ./${loader}/credentials.py
+
+update ${download} ${dataset} 
+
+if [ "${success}" = "no" ]; then
+	echo "loading ended"
+	exit 1
+fi
 
 eval "$(conda shell.bash hook)"
 cd ${release}
 conda activate standard
 echo "starting load `pwd`"
 date
-time python -u ../RMCLoader/readrmcfiles.py
+time python -u ../${loader}/readrmcfiles.py
 
 cd ..
 
 del() {
   shift
   for d in $*; do
-        echo 'removing dataset' $d
-        rm -r "${d}"
+	if [ ! "${d}" == "${release}" ]; then
+          echo 'removing dataset' $d
+          rm -r "${d}"
+	fi
   done
 
 }
